Break fields button into a separate component
