+++ react/components/TableCompose/ToolbarInput.js
@@ -3,9 +3,13 @@ import PropTypes from 'prop-types'
 
 import InputSearch from '../InputSearch'
 
-const ToolbarInput = ({ onSubmit, disabled, inputSearch }) => {
+const ToolbarInput = ({ disabled, inputSearch }) => {
+  const handleInputSearchSubmit = e => {
+    !!inputSearch.onSubmit && inputSearch.onSubmit(e)
+  }
+
   return (
-    <form className="w-40" onSubmit={onSubmit}>
+    <form className="w-40" onSubmit={handleInputSearchSubmit}>
       <InputSearch disabled={disabled} {...inputSearch} />
     </form>
   )
@@ -16,7 +20,6 @@ ToolbarInput.propTypes = {
     onSubmit: PropTypes.func,
   }),
   disabled: PropTypes.bool,
-  onSubmit: PropTypes.func,
 }
 
 export default ToolbarInput
